implement SafeArchiveInputStream #3029
